\name{internal-functions}
\alias{em_damda}
\alias{init_em}
\alias{mstep_damda}
\alias{start_parallel}
\alias{bic_reg}
\alias{ice}
\alias{estepdamda}
\alias{icf}
\alias{bhat_dist}

\title{
Internal \code{damda} function
}
\description{
Internal function not to be called by the user.
}
\value{
No return value, these functions are not meant to be called by the user.
}

\keyword{internal}
